User validation implemented and file naming convention changed

+++ src/server/api/routers/Files/delete-file.ts
@@ -13,6 +13,17 @@ export const deleteFile = publicProcedure
     const { fileId } = input;
 
     try {
+      const file = await db.folder.findFirst({
+        where: {
+          id: fileId,
+        },
+      });
+      if(!file || file.userId !== ctx?.userId){
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          message: "You are not authorized to delete this file",
+        });
+      }
       const deletedFile = await db.file.delete({
         where: {
           id: fileId,
